feat: add condition to check and allow creating account on wallet dropdown
